Refactor tokens list module structure (#1863)
