+++ src/cow-react/modules/application/containers/App/Updaters.tsx
@@ -20,7 +20,7 @@ import {
 import { UploadToIpfsUpdater } from 'state/appData/updater'
 import { GasPriceStrategyUpdater } from 'state/gas/gas-price-strategy-updater'
 import { EthFlowSlippageUpdater, EthFlowDeadlineUpdater } from '@cow/modules/swap/state/EthFlow/updaters'
-import { TokensListUpdater } from '@cow/modules/tokensList/TokensListUpdater'
+import { TokensListUpdater } from '@cow/modules/tokensList/updaters/TokensListUpdater'
 
 export function Updaters() {
   return (
